Automatic Deployment of our CSV Database Web Service has been added
Changed URI to our group web database, so we automatic deployment of our CSV database Web Service will be added. Beside that 4 lines in our GitHub Actions workflow that was added to our repository has been changed, because our project is organized into src and test directories, and so that it can be build and publish from the correct directory.

Co-Authored-By: Johan Ingeholm <151632146+JIngeholm@users.noreply.github.com>
Co-Authored-By: JacquesPuvis <147432028+JacquesPuvis@users.noreply.github.com>
Co-Authored-By: Noaleflame <151853768+Noaleflame@users.noreply.github.com>
Co-Authored-By: Freya Nørlund Wentzel <145338621+fr3yaw3ntz3l@users.noreply.github.com>