UITests and a custom web application for tests has been made
The browser and context is disposed after each test, so we make sure that we dont make "testusers" directly in our program's chirp, but so that we will have a database with testusers, which then are being disposed after each test.

Right now it is only the first two tests that can run, since in the other tests the register and login part need to be inserted, because with the custom web application for tests, the users registered in one test are not accessible in another.
Suggestion to change: create a setup for the registration and login part, create one long test?