All tests are working with CustomTestWebApplication
Fixes Playwright test suite, and test are running on a random port to ensure that all test can run on the same time, and we dont get an "Address already in use" error.
Also making sure that the authenticationscheme is the Identity:Application, so that we can use the SignInManager property, to for example login.
Also making sure to tear down each of the browsers after each test, and also sets up register and login for tests that doesnt test the register and login page.