UnitTests for CheepRepository
Realized that AuthorDTOs  as an object wasn't being used, so I changed name in CheepDTO since its the authors name that the field holds. Changed CheepDBContext and migrations since we realized that the cheep's authorid might not have been referenced as a foreign key to the author's id entirely correct.
Test coverage of CheepRepository is now 100%.
Removed delays in UITests to have them run faster for workflow.
Moved null check in SaveCheep as first, to insure thats never the case and to check it in tests.

Co-authered-by: chatGPT